[user] update form for create and update users
