+++ src/components/header/styles.scss
@@ -13,7 +13,7 @@
   position: absolute;
   top:$top-menu;
   display:grid;
-  grid-template-columns: 10% 90%;
+  grid-template-columns: 10% 80% 10%;
   width:100%;
 }
