+++ src/components/styles/Global.scss
@@ -35,6 +35,7 @@ html {
     background-color: $background-color-form;
     padding: 10px;
     border-radius: 0 $border-principal;
+    margin-bottom: 10px;
   }
 
   .form-section__title{
@@ -53,13 +54,14 @@ html {
     padding: 10px 5px 5px 5px;
     margin-bottom: 20px;
     outline: none;
+    font-size: $font-size-principal;
   }
   
   .form-section__buttons{
     display: flex;
     flex-direction: row;
     justify-content: space-around;
-    margin-top: 20px;
+    margin-top: 15px;
   }
   
   .form-button{
